Update normalize.js
Revert treatment of 'ph' so it can be dealt with separately in general-use mode.  Add mapping of apostrophe key to forward slash to allow non-technical users to be able to type an apostrophe or forward slash on the keyboard and have it render an 'apostrophe like' character and not error out.